Displacement function for shapekey mode in wrinkle maps.
NOTE: this is not very useful atm, because shapekeys are always in
object space. In order for shape keys to follow the mesh deformation
they have to be applied *before* and armature or other deformations
as a virtual modifier!

With the current single modifier pass this is not possible. A potential
workaround could be to run a second modifier pass up to the wrinkle modifier.
The wrinkle mod would first take the base DM input without the wrinkle shapes,
measure the triangle deformations, then re-eval the modifier stack up to
the wrinkle mod with the correctly weighted shapes applied.
This is not a very nice solution and might cause unforeseen issues.