Revert "Wrinkle map type, to select between shapekey or texture input."
This reverts commit 3856ad8afa9ee42fe1893046f68cafb74cf71f3f.

We can use an additional option in shapekeys to designate wrinkle maps,
rather than keep a complicated extra list of weak references in the modifier.

It would be nice to do the same for texture maps, but there is no equivalent
object-scope list of textures (apart from the material lists which have a
different purpose and don't even exist in this form in Cycles any more).