Make the wrinkle modifier require original topology to make shapekey eval easier.
The deformation of triangles could be measured for subdivided or otherwise
modified topology as well. But in order to generate vertex groups for shapekeys
the modifier would have to map back from the hires DM, and then sum up
the influence of all relevant triangles. This is difficult and unnecessary,
since the modifier really only makes sense when applied on the base geometry.